# OAS/oas-server/api/oas-doc.yaml
openapi: 3.0.0
info:
  title: School API
  version: '1.0'
  description: Get Basic Academic Information Through School API
paths:
  /api/students/:
    get:
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: number
                    Name:
                      type: string
                    Address:
                      type: string
                    Courses:
                      type: array
                      items:
                        type: object
                        properties:
                          id:
                            type: string
                          grade:
                            type: string
                          name:
                            type: string
      summary: Get all students infomations
    post:
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: number
                    name:
                      type: string
                    address:
                      type: string
                    courses:
                      type: array
                      items:
                        type: object
                        properties:
                          id:
                            type: number
                          grade:
                            type: number
                          name:
                            type: string
              examples:
                response:
                  value:
                    - id: 1
                      name: Veli-Matti Lilja
                      address: Norra Larsmovägen 38 70820 Oulu
                      courses:
                        - id: 1
                          grade: 5
                          name: Php
                        - id: 2
                          grade: 3
                          name: Hybrid Application
                        - id: 3
                          grade: 4
                          name: ReactJS
                    - id: 2
                      name: Erno Virolainen
                      address: Kankaanpääntie 36 94600 KEMI
                      courses:
                        - id: 1
                          grade: 3
                          name: Php
                        - id: 3
                          grade: 3
                          name: ReactJS
                    - id: 3
                      name: Kaarina Autio
                      address: Ysitie 25 33420 TAMPERE
                      courses:
                        - id: 2
                          grade: 0
                          name: Hybrid Application
                        - id: 3
                          grade: 2
                          name: ReactJS
        '400':
          description: ''
      summary: Add a new student
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: number
                address:
                  type: string
                name:
                  type: string
                courses:
                  type: object
                  properties:
                    id:
                      type: number
                    grade:
                      type: number
                    name:
                      type: string
        x-examples:
          application/json:
            id: 3
            name: Kaarina Autio
            address: Ysitie 25 33420 TAMPERE
            courses:
              - id: 2
                grade: 0
                name: Hybrid Application
              - id: 3
                grade: 2
                name: ReactJS
  '/api/students/{students_id}':
    put:
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties: {}
        '400':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: number
                    name:
                      type: string
                    address:
                      type: string
                    courses:
                      type: array
                      items:
                        type: object
                        properties:
                          id:
                            type: number
                          grade:
                            type: number
                          name:
                            type: string
              examples:
                response:
                  value:
                    - Id: 1
                      name: Veli-Matti Lilja
                      Address: Norra Larsmovägen 38 70820 Oulu
                      Courses:
                        - id: 1
                          grade: 5
                          name: Php
                        - id: 2
                          grade: 3
                          name: Hybrid Application
                        - id: 3
                          grade: 4
                          name: ReactJS
                    - id: 2
                      name: Erno Virolainen
                      address: Kankaanpääntie 36 94600 KEMI
                      courses:
                        - id: 1
                          grade: 3
                          name: Php
                        - id: 3
                          grade: 3
                          name: ReactJS
                    - id: 3
                      name: Kaarina
                      address: Ysitie 25 TAMPERE
                      courses:
                        - id: 2
                          grade: 4
                          name: Hybrid Application
                        - id: 3
                          grade: 5
                          name: ReactJS
        '404':
          description: ''
      summary: Update Student by id
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: number
                  name:
                    type: string
                  address:
                    type: string
                  courses:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: number
                        grade:
                          type: number
                        name:
                          type: string
        x-examples:
          application/json:
            id: 3
            name: Kaarina
            address: Ysitie 25 TAMPERE
            courses:
              - id: 2
                grade: 4
                name: Hybrid Application
              - id: 3
                grade: 5
                name: ReactJS
    get:
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties: {}
        '404':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: number
                    name:
                      type: string
                    address:
                      type: string
                    courses:
                      type: array
                      items:
                        type: object
                        properties:
                          id:
                            type: number
                          grade:
                            type: number
                          name:
                            type: string
              examples:
                response:
                  value:
                    - id: 1
                      name: Veli-Matti Lilja
                      address: Norra Larsmovägen 38 70820 Oulu
                      courses:
                        - id: 1
                          grade: 5
                          name: Php
                        - id: 2
                          grade: 3
                          name: Hybrid Application
                        - id: 3
                          grade: 4
                          name: ReactJS
                    - id: 3
                      name: Kaarina Autio
                      address: Ysitie 25 33420 TAMPERE
                      courses:
                        - id: 2
                          grade: 0
                          name: Hybrid Application
                        - id: 3
                          grade: 2
                          name: ReactJS
      summary: Get Student By id
    delete:
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties: {}
        '401':
          description: ''
        '404':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: number
                    name:
                      type: string
                    address:
                      type: string
                    courses:
                      type: array
                      items:
                        type: object
                        properties:
                          id:
                            type: number
                          grade:
                            type: number
                          name:
                            type: string
              examples:
                response:
                  value:
                    - Id: 1
                      name: Veli-Matti Lilja
                      Address: Norra Larsmovägen 38 70820 Oulu
                      Courses:
                        - id: 1
                          grade: 5
                          name: Php
                        - id: 2
                          grade: 3
                          name: Hybrid Application
                        - id: 3
                          grade: 4
                          name: ReactJS
      summary: Delete A Student By id
      parameters:
        - in: path
          name: students_id
          required: true
          schema:
            type: number
            default: '2'
...
servers:
  - url: 'http://shoolExample.com'
  - url: 'https://shoolExample.com'